About Page/Removed Extra Features/Updated API Code and Apperance
Added an About page. Removed settings as it is not needed for a MVP. Updated the way the API is being handled. Improved the general appearance of the application.

+++ app-dev-practice/src/app/show-summary-table/show-initial-form/show-initial-form.component.ts
@@ -1,7 +1,10 @@
-import { Component } from '@angular/core';
+import { Component, Input } from '@angular/core';
 import { FormBuilder, Validators } from '@angular/forms';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { MatDialog, MatDialogRef } from '@angular/material/dialog';
+import { Show } from 'src/app/interfaces/show.interface';
+import { Inject } from '@angular/core';
+import { MAT_DIALOG_DATA } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-show-initial-form',
@@ -16,7 +19,8 @@ export class ShowInitialFormComponent {
   constructor(
     private formBuilder: FormBuilder,
     private http: HttpClient,
-    public dialogRef: MatDialogRef<ShowInitialFormComponent>
+    public dialogRef: MatDialogRef<ShowInitialFormComponent>,
+    @Inject(MAT_DIALOG_DATA) public existingShows: Show[]
   ) {
     this.showDetailsForm = this.formBuilder.group({
       showName: [''],
@@ -39,8 +43,16 @@ export class ShowInitialFormComponent {
       next: (showName: any) => {
         this.searchShow(showName).subscribe({
           next: (response: any) => {
-            localStorage.setItem('showDetails', JSON.stringify(response));
-            this.lockNext = false;
+            let showAlreadyExists = false;
+            showAlreadyExists =
+              Array.isArray(this.existingShows) &&
+              this.existingShows.some(
+                (existingShow) => existingShow.showName === response['showName']
+              );
+            if (!showAlreadyExists) {
+              localStorage.setItem('showDetails', JSON.stringify(response));
+              this.lockNext = false;
+            }
           },
           error: (error: HttpErrorResponse) => {
             this.errors = error.error;
